+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -5,6 +5,8 @@ Release History
 (unreleased)
 +++++++++++++++++++
 * webapp: fix the bug that listing linux webapp returns nothing
+* webapp: setting runtime is mandatory for linux
+* webapp: use only linux-fx-version instead of CUSTOM_DOCKER_IMAGE_NAME
 
 0.1.10 (2017-07-07)
 +++++++++++++++++++